Change loader placement in card
